pcache/client/redis: support multi-threaded environments
With this change, the Redis implementation is now safe to be used in a
multi-threaded environment as it uses a JedisPool that is a connection
pool to Redis.

Now each call to Redis starts with retrieving of connection from that
pool and releasing it after the operation completes.

More information for multi-threaded usage could be found at https://github.com/xetorthio/jedis/wiki/Getting-started